fix(r/primary_analysis.r): change control SD to population SD
on debugging, I realized that the telomere and control Ct SD calculations are using sample and
population SD respectively. I'm not sure why this is and will ask for clarification. But until that
time, I've added support for population SD and appropriate documentation, to more closely emulate
the results in Data/Analysis/*xlsx.