fix(r/primary_analysis.r): add partial, pretty broken content for dynamic well mapping
I've been trying to fold in the external plate list and contents files, and it's just been a slog.
I'm not sure exactly how to recompute the output from the template without assuming a fixed well
mapping (from 96 to 384); it's possible it's actually impossible and they really do require a fixed
layout, at least for the standards and single stock NTC. I'll take some time to reflect on it. See
inline comments for more info.

re #6 #12